requirements.txt
# ── LLM ──────────────────────────────────────────────────────────────────────
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0

# ── Notion ────────────────────────────────────────────────────────────────────
notion-client>=2.2.1

# ── GitHub 클론 ───────────────────────────────────────────────────────────────
gitpython>=3.1.43

# ── UI ───────────────────────────────────────────────────────────────────────
streamlit>=1.35.0

# ── 환경변수 ──────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1

# ── 로깅 ─────────────────────────────────────────────────────────────────────
pyyaml>=6.0.1